ddi_app/requirements.txt
streamlit
langgraph>=0.2.34
neo4j>=5.20.0
pandas
python-dotenv
httpx
filelock
openai>=1.40.0
pydantic>=2.7.0
beautifulsoup4
lxml
tenacity
aiosqlite

